Add Netlify Redirect URL
To get Netlify to work with the correct URLs and still be able to use GitHub Pages I add a redirect url.
For now I want to still be able to use GitHub Pages because some (important) links still link to it. This is the easiest way to do just that.