add .yarnrc for netlify
based on this: https://answers.netlify.com/t/using-the-new-yarn-release-2-0-0-berry/8270/5
"I managed to get it work aftet renaming the .yarnrc.yml configuration file to .yarnrc. Berry started malfunctioning, and I learned that .yarnrc is NOT equivalent to .yarnrc.yml. Yet, Netlify still requires the former to enable v2. So we need both an .yarnrc file and a .yarnrc.yml file."